Implement tests for main functions (#2)
* [tests] Implement testing for main functions

Implements tests for transcription and diarization functions to prevent
regressions. The audio sample used in these firsts tests was chosen for
its inclusion of historically minoritized speakers, conversational
nature similar to sociolinguistic interviews, and overlapping
speech/back-channeling as an early benchmark for dealing with
challenges of naturalistic data.

* [workflows] Use dev and prod requirements.txt

Put dev requirements in their own requirements-dev.txt which calls the
main requirements.txt and use that in the workflow. Previously, all the
dev requirements were listed manually in an install call which was bad
for portability and made us call the install script twice.