[formatter] Implement conversion to TextGrid (#4) (#8)
* [formatter] Implement conversion to TextGrid (#4)

Introduce a class which houses functions to convert the json output of
the transcriber and diarizer into formats that can be input into various
alignment software. This commitn introduces a conversion to TextGrids
for use with MFA though other formats may be introduced later.

* [tests] Only test transcript equivalence (#5)

Tests of alignment and speaker assignment no longer test for equivalence
of start and end times or assignment weights because different machines
have slightly different results leading to failures. Instead, only the
transcriptions themselves are asserted to be equivalent.

Fixes #5

* [formatter] Implement conversion to TextGrid (#4)

Introduce a class which houses functions to convert the json output of
the transcriber and diarizer into formats that can be input into various
alignment software. This commitn introduces a conversion to TextGrids
for use with MFA though other formats may be introduced later.

* Avoid gaps between phrases

Lots of work arounds to handle the gaps between phrases in the
timestamps, but after looking at them, they should probably just be
included in the preceding segment because they tend to contain a fair
amount of speech.

* Revert "Avoid gaps between phrases"

This reverts commit 54c0c59c20c53254f1e550bdbebb518d6c143b02.